+++ packages/core/src/table/hooks/use_crush_all_cols.ts
@@ -1,11 +1,12 @@
 import React from "react";
 import { useMeasureContext } from "./use_measure_context";
+import { CrushBy } from "../types";
 
 export const useCrushAllCols = () => {
   const { crushAllColumns } = useMeasureContext();
   const crushAllColumnsRef = React.useRef(crushAllColumns);
   crushAllColumnsRef.current = crushAllColumns;
-  return React.useCallback(() => {
-    crushAllColumnsRef.current();
+  return React.useCallback((crushBy: CrushBy | "default" = "default") => {
+    crushAllColumnsRef.current(crushBy);
   }, []);
 };