+++ packages/core/src/table/table_context.tsx
@@ -1,13 +1,18 @@
 "use client";
+import { ColumnDefTemplate, HeaderContext } from "@tanstack/react-table";
 import React from "react";
-import { RttuiTable } from "./types";
 import { Skin } from "../skin";
+import { RttuiTable } from "./types";
 
 export type TableContextType = {
   tableContainerRef: React.RefObject<HTMLDivElement | null>;
   skin: Skin;
   loading: boolean;
   tableRef: React.RefObject<RttuiTable>;
+  renderHeaderPlaceholder: (
+    headerDef: undefined | ColumnDefTemplate<HeaderContext<any, any>>,
+    headerContext: HeaderContext<any, any>,
+  ) => React.ReactNode;
 };
 
 export const TableContext = React.createContext<TableContextType | undefined>(