+++ packages/skin-bleu/src/filter.tsx
@@ -1,8 +1,9 @@
-import { Input } from "@mui/material";
+import { Box, Input } from "@mui/material";
 import { shallowEqual, useColProps, useColRef } from "@rttui/core";
 import React from "react";
+import { Resizer } from "./resizer";
 
-export function Filter() {
+export function Filter({ resizer }: { resizer?: boolean }) {
   const { filterValue, canFilter } = useColProps({
     callback: ({ column }) => {
       return {
@@ -15,23 +16,35 @@ export function Filter() {
   });
   const colRef = useColRef();
 
-  if (!canFilter) {
+  if (!canFilter && !resizer) {
     return null;
   }
 
   return (
-    <Input
-      size="small"
-      fullWidth
+    <Box
       sx={{
-        minWidth: "120px",
+        display: "flex",
+        alignItems: "center",
+        flex: 1,
       }}
-      value={filterValue}
-      onChange={(e) => {
-        React.startTransition(() => {
-          colRef().column?.setFilterValue(e.target.value);
-        });
-      }}
-    />
+    >
+      {canFilter && (
+        <Input
+          size="small"
+          fullWidth
+          sx={{
+            minWidth: "120px",
+          }}
+          value={filterValue}
+          onChange={(e) => {
+            React.startTransition(() => {
+              colRef().column?.setFilterValue(e.target.value);
+            });
+          }}
+        />
+      )}
+      {resizer && <Box sx={{ flex: 1 }} />}
+      {resizer && <Resizer />}
+    </Box>
   );
 }