+++ nextjs-test/CHANGELOG.md
@@ -1,5 +1,12 @@
 # nextjs-test
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @rttui/core@1.0.47
+
 ## 0.1.2
 
 ### Patch Changes